declared dep check in the wrong place, now corrected
